Refactor implementation, tests and samples to support new module name
