+++ native/src/main/java/org/ballerinax/azure/sb/util/ASBConstants.java
@@ -16,7 +16,7 @@
  * under the License.
  */
 
-package org.ballerinax.azure.asb.util;
+package org.ballerinax.azure.sb.util;
 
 import io.ballerina.runtime.api.utils.StringUtils;
 import io.ballerina.runtime.api.values.BString;
